requirements.txt
Base
streamlit>=1.34
langchain>=0.2
openai>=1.30
faiss-cpu>=1.7
tiktoken>=0.7
pydantic>=2.7
python-dotenv>=1.0
pandas>=2.2

# For multi-format ingestion & chains
langchain[pdf]>=0.2
unstructured>=0.7
python-pptx>=0.7  # later slide export